feat(localizer): add FluentLocalizerBuilder for flexible localizer construction
Introduce `FluentLocalizerBuilder` to enable customized construction of the `FluentLocalizer`. This builder supports specifying the locale, overlaying custom Fluent resources or a pre-built consumer bundle, disabling default resource loading, and configuring an error reporter for formatting issues. The builder validates locale consistency and resource parsing when building the localizer, improving extensibility and correctness of localization setup.
